+++ network-api/networkapi/wagtailpages/tests/test_utils.py
@@ -16,19 +16,12 @@ from networkapi.wagtailpages import (
     to_language,
 )
 from networkapi.wagtailpages.factory import blog as blog_factories
-from networkapi.wagtailpages.factory.libraries.research_hub import (
-    detail_page as detail_page_factory,
-)
-from networkapi.wagtailpages.factory.libraries.research_hub import (
-    relations as relations_factory,
-)
 from networkapi.wagtailpages.factory.profiles import ProfileFactory
 from networkapi.wagtailpages.pagemodels.blog.blog import BlogAuthors
 from networkapi.wagtailpages.pagemodels.profiles import Profile
 from networkapi.wagtailpages.utils import (
     create_wagtail_image,
     get_blog_authors,
-    get_research_authors,
     localize_queryset,
 )
 
@@ -265,41 +258,6 @@ class TestLocalizeQueryset(TestCase):
         # Assert that the items are ordered by the name in the active locale
         self.assertEqual(result[0], apple)
         self.assertEqual(result[1], banana)
-
-
-class TestGetResearchAuthors(TestCase):
-    def test_get_research_authors(self):
-        research_author_profile = ProfileFactory()
-        relations_factory.ResearchAuthorRelationFactory(
-            detail_page=detail_page_factory.ResearchDetailPageFactory(),
-            author_profile=research_author_profile,
-        )
-        not_research_author_profile = ProfileFactory()
-
-        research_author_profiles = get_research_authors(Profile.objects.all())
-
-        self.assertIn(research_author_profile, research_author_profiles)
-        self.assertNotIn(not_research_author_profile, research_author_profiles)
-
-    def test_get_research_authors_distinct(self):
-        """Return research author profile only once"""
-
-        research_author_profile = ProfileFactory()
-        relations_factory.ResearchAuthorRelationFactory(
-            detail_page=detail_page_factory.ResearchDetailPageFactory(),
-            author_profile=research_author_profile,
-        )
-        relations_factory.ResearchAuthorRelationFactory(
-            detail_page=detail_page_factory.ResearchDetailPageFactory(),
-            author_profile=research_author_profile,
-        )
-
-        profiles = Profile.objects.all()
-        profiles = get_research_authors(profiles)
-        profiles = profiles.filter(id=research_author_profile.id)
-        count = profiles.count()
-
-        self.assertEqual(count, 1)
 
 
 class TestGetBlogAuthors(TestCase):
